Refactor: Separate Profile.jsx component - Profile 컴포넌트를 jsx 파일로 분리
