+++ src/pages/SignUp/index.jsx
@@ -1,9 +1,9 @@
 import styled from "styled-components";
 
-import CameraIcon from "../../assets/CameraIcon.svg";
 import Header from "./Header";
 
 import { useState } from "react";
+import Profile from "./Profile";
 
 export default function SignUp() {
   // 핸드폰 번호
@@ -50,28 +50,6 @@ export default function SignUp() {
     } else {
       setPasswordErrorMsg("");
       setPasswordConfirm(e.target.value); // 실제 넘길 값
-    }
-  };
-
-  // 프로필
-  const handleImageChange = (e) => {
-    const file = e.target.files[0];
-
-    if (file) {
-      if (file.size > 5 * 1024 * 1024) {
-        alert("이미지 크기는 5MB를 초과할 수 없습니다.");
-      } else {
-        const allowedExtensions = ["jpg", "jpeg", "png"];
-        const fileExtension = file.name.split(".").pop().toLowerCase();
-
-        if (allowedExtensions.includes(fileExtension)) {
-          setProfileImage(file);
-        } else {
-          alert(
-            "지원되지 않는 형식입니다. jpg, jpeg, png 이미지 형식을 사용해주세요."
-          );
-        }
-      }
     }
   };
 
@@ -159,20 +137,7 @@ export default function SignUp() {
       <SignUpContainer>
         <Header />
         <Form>
-          <ProfileWrapper>
-            <Profile
-              htmlFor="profileImageInput"
-              imageUrl={profileImage && URL.createObjectURL(profileImage)}
-            >
-              <img src={CameraIcon} alt="Camera Icon" />
-              <input
-                type="file"
-                id="profileImageInput"
-                accept=".jpg, .jpeg, .png"
-                onChange={handleImageChange}
-              />
-            </Profile>
-          </ProfileWrapper>
+          <Profile />
           <InputWrapper>
             <InputTitle>
               닉네임<span>*</span>
@@ -272,47 +237,6 @@ const SignUpContainer = styled.div`
   display: flex;
   flex-direction: column;
   align-items: center;
-`;
-
-const ProfileWrapper = styled.div`
-  display: flex;
-  justify-content: center;
-`;
-
-const Profile = styled.label`
-  width: 104px;
-  height: 104px;
-  background-color: #d9d9d9;
-  border-radius: 50%;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  cursor: pointer;
-  overflow: hidden;
-  position: relative;
-
-  input {
-    display: none;
-  }
-  img {
-    width: 32px;
-    height: 32px;
-    cursor: pointer;
-  }
-
-  &:before {
-    content: "";
-    width: 100%;
-    height: 100%;
-    background-image: url(${(props) => props.imageUrl});
-    background-size: cover;
-    background-position: center;
-    border-radius: 50%;
-    position: absolute;
-    top: 0;
-    left: 0;
-    cursor: pointer;
-  }
 `;
 
 const Form = styled.form`
